# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API documentation for the Library application",
        "version": "1.0.0"
    },
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "List all authors",
                "description": "Returns a list of authors",
                "operationId": "33b4ca0e528a8b6298d9942781a0c85a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Create a new author",
                "description": "Creates a new author",
                "operationId": "d998cf74685b2e9609bcd48ab986f678",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get an author by ID",
                "description": "Returns a single author",
                "operationId": "5317eb225dd93ee043c24cfacda9f26d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update an existing author",
                "description": "Updates an existing author",
                "operationId": "54818ef139366ae08ef5747f7c9d9a36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Author updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Author not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete an author",
                "description": "Deletes an author by ID",
                "operationId": "c8d794bbc8f06854b38ce720c4395db8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Author deleted"
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "List all books",
                "description": "Returns a list of books",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Creates a new book",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a book by ID",
                "description": "Returns a single book",
                "operationId": "e94109dcb4f4d40b58a210b43d94016d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update an existing book",
                "description": "Updates an existing book",
                "operationId": "7430c1078c3cb5da552b2fd50409f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "description": "Deletes a book by ID",
                "operationId": "41c44a46960ad1a921ba5bfbf16ea7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
        "/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "List all loans",
                "description": "Returns a list of loans",
                "operationId": "7a6d018dd0b622136235484db1036de8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Loan"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Create a new loan",
                "description": "Creates a new loan",
                "operationId": "1f6c6e0db1616c54e249195c256856c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Loan"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/loans/{id}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Get a loan by ID",
                "description": "Returns a single loan",
                "operationId": "a4a657443e9563396f74613aa1059203",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the loan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Loans"
                ],
                "summary": "Update an existing loan",
                "description": "Updates an existing loan",
                "operationId": "9851fce80ce3bb9682a7ae6d34fda306",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the loan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Loan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loan updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loan not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Loans"
                ],
                "summary": "Delete a loan",
                "description": "Deletes a loan by ID",
                "operationId": "3aaa9683b74a4086d39b53b57c0598d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the loan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Loan deleted"
                    },
                    "404": {
                        "description": "Loan not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "properties": {
                    "id": {
                        "description": "ID of the author",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the author",
                        "type": "string"
                    },
                    "birth_date": {
                        "description": "Birth date of the author",
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "description": "Timestamp when the author was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the author was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Book": {
                "properties": {
                    "id": {
                        "description": "ID of the book",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string"
                    },
                    "publication_year": {
                        "description": "Publication year of the book",
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "description": "Timestamp when the book was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the book was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "authors": {
                        "description": "List of authors associated with the book",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}